requirements.txt
pyyaml>=6.0
click>=8.0
jinja2>=3.1
python-docx>=1.1
